refactor: Sign Up Page (#5)
