+++ frontend/package.json
@@ -18,17 +18,22 @@
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^4.1.3",
     "@tailwindcss/vite": "^4.0.14",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "lucide-react": "^0.483.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
     "react-router-dom": "^7.4.0",
     "react-slideshow-image": "^4.3.2",
+    "tailwind-merge": "^3.2.0",
     "tailwindcss": "^4.0.14",
+    "tw-animate-css": "^1.2.5",
     "zod": "^3.24.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
+    "@types/node": "^22.14.0",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
     "@vitejs/plugin-react-swc": "^3.8.0",
