+++ frontend/src/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import LogInPage from "./pages/LogInPage.jsx";
-import SignUpPage from "./pages/SignUp/SignUpPage.jsx";
+import SignUpPage from "./pages/SignUp/components/SignUpPage.jsx";
 import HomePage from "./pages/Home/HomePage.jsx";
 import DashboardPage from "./pages/Dashboard/DashboardPage.jsx";
 import AddFoodPage from "./pages/AddFood/AddFoodPage.jsx";
